requirements.txt
discord.py==2.5.2
yt-dlp==2025.6.30
